Refactor code to clean up main()
